[FIX] website_slides: use email_from on new content notif
- Seta a value for the `ir.config_parameter` `mail.catchall.domain`
- Enroll user A to a course
- Add new content to the course
- Publish it
  => an email is sent to the users enrolled
- Reply to the email

The reply is considered as a review of the course.

It is not intended that users reply to such email; they are 'one-way'
notifications.

A solution is to be able to set the `reply_to` field on the mail
template. This way, it's possible to set it to a `noreply` value.

opw-2290521

closes odoo/odoo#54281

X-original-commit: 2f063fa4ce8d051858419aba6900496ad0358833
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>