[FIX] web: remove test-breaking css property
This is a huge mistery. But this particular `align-items: baseline`
rule, introduced by [1], yields a red runbot ~20% of the time.

As it is not stricly necessary in current master (but necessary if the
font changes which will be the case in the upcoming days), this commit
simply removes the rule temporarily.

[1]: https://github.com/odoo/odoo/commit/5f7cb675743fbb094edb4c4ea682fc6476c09c80

Part of https://github.com/odoo/odoo/pull/54274

closes odoo/odoo#54274

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>