[FIX] website: remove duplicated filter option for blockquote
The blockquote in its "cover" layout allows to add a background with
a dedicated colored filter on top of it. That option was a bit buggy
since the filter option only appeared if the page was saved first. As
we now have a generic option to apply a colored filter on image
themselves, that option became useless anyway.

Part of https://github.com/odoo/odoo/pull/53930
task-2157252