[FIX] point_of_sale: Stop led_status.sh on update
The fact that led_status.sh was still running when performing the
`git reset --hard` made it impossible to remount the filesystem in RO
mode, due to a `mount point is busy` error.

Instead of starting the led_status_sh script from rc.local, we make it
a service so it can be easily restarted when an update is performed.

closes odoo/odoo#54353

X-original-commit: 2ad99d3b9381a61e112207c79ac89cdf4ebc6c3c
Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>
Signed-off-by: Antoine Prieëls <aprieels@users.noreply.github.com>