[FIX] base, mail: properly annotate Partner.find_or_create
In 4a54cb60c8020ca84a693a11d61bce8c880b83e7 `find_or_create` was
modified to return a recordset rather than just an id. This made it
incompatible with the RPC protocol and it was not annotated to
downgrade the return value back to an id.

Fixes odoo/odoo#54466

closes odoo/odoo#54476

X-original-commit: ad018b52d1ea2cf7c7f277940dc14c351b7f22f5
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>