[FIX] point_of_sale,pos_cash_rounding: return payment with rounding
In the PoS, activate the cash rounding.
Open a PoS session and make a sale.
Close the session (without validating the entries yet),
Go to the just made order to return the items.

Upon making return payment, it gets stuck, because Odoo offers
to refund the unrounded amount and it doesn't match what the customer
paid.

opw-2289751

closes odoo/odoo#54870

X-original-commit: 6d6e7d9247b47e0c8e8b735177e6385b28eb874a
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: agr-odoo <agr-odoo@users.noreply.github.com>