[IMP] hr_timesheet: Add argument so method can be override by other module
Currently, method _apply_timesheet_label will add label on field
unit_amount with widget timesheet_uom but the widget on this
field is changed in another module as timesheet_uom_timer so it will
not work when that module is installed.

So in this method pass the view_type as argument so it can be easily
override in another module based on view_type.

closes odoo/odoo#57847

Taskid: 2287390
Related: https://github.com/odoo/enterprise/pull/11884
Related: odoo/enterprise#13279
X-original-commit: 7754ddf71706762cd79c53acacb5dbb485bcf3f2
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>